INFRA-9248 - Fix for docker container restart fail

+++ docker/docker-entrypoint.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env sh
 set -eu
 
-envsubst '${API_HOST}' < /etc/nginx/conf.d/default.conf.template > /etc/nginx/nginx.conf 
-
-rm /etc/nginx/conf.d/default.conf.template 
+if test -f /etc/nginx/conf.d/default.conf.template  ; then
+    envsubst '${API_HOST}' < /etc/nginx/conf.d/default.conf.template > /etc/nginx/nginx.conf 
+    rm /etc/nginx/conf.d/default.conf.template 
+fi
 
 exec "$@"
